# apps/api/docs/swagger.yaml
definitions:
...
  document.DocumentResponse:
    properties:
      checksum:
        example: abc123def456
        type: string
      company_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      created_at:
        example: "2024-11-28T15:04:05Z"
        type: string
      download_url:
        example: https://minio.example.com/bucket/document.pdf?token=...
        type: string
      file_path:
        example: documents/550e8400-e29b-41d4-a716-446655440000.pdf
        type: string
      file_size:
        example: 1024000
        type: integer
      folder_id:
        example: 550e8400-e29b-41d4-a716-446655440002
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      mime_type:
        example: application/pdf
        type: string
      preview_file_path:
        example: previews/550e8400-e29b-41d4-a716-446655440000.jpg
        type: string
      preview_url:
        example: https://minio.example.com/bucket/preview.jpg?token=...
        type: string
      sender_id:
        example: 550e8400-e29b-41d4-a716-446655440003
        type: string
      tags:
        items:
          $ref: '#/definitions/document.TagData'
        type: array
    type: object
  document.DocumentsListResponse:
    properties:
      documents:
        items:
          $ref: '#/definitions/document.DocumentResponse'
        type: array
      total:
        example: 10
        type: integer
    type: object
  document.SearchRequest:
    properties:
      company_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      folder_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      page:
        example: 1
        type: integer
      page_size:
        example: 20
        type: integer
      query:
        example: invoice
        type: string
      tag_ids:
        items:
          type: string
        type: array
    required:
    - company_id
    type: object
  document.TagData:
    properties:
      company_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: Important
        type: string
    type: object
...
  document.UpdateRequest:
    properties:
      folder_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: new_name.pdf
        minLength: 1
        type: string
      sender_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
    type: object
...
  documenttag.TagResponse:
    properties:
      company_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: Important
        type: string
    type: object
...
  documenttag.UpdateTagRequest:
    properties:
      name:
        example: Very Important
        minLength: 1
        type: string
    required:
    - name
    type: object
...
  folder.FolderResponse:
    properties:
      company_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      count:
        example: 0
        type: integer
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: Documents
        type: string
      parent_folder_id:
        example: 550e8400-e29b-41d4-a716-446655440002
        type: string
      size:
        example: 0
        type: integer
    type: object
  folder.FoldersListResponse:
    properties:
      folders:
        items:
          $ref: '#/definitions/folder.FolderResponse'
        type: array
      total:
        example: 10
        type: integer
    type: object
  folder.GetByParentRequest:
    properties:
      company_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      parent_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
    required:
    - company_id
    type: object
  folder.RenameRequest:
    properties:
      name:
        example: New Folder Name
        minLength: 1
        type: string
    required:
    - name
    type: object
  handlers.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
info:
  contact: {}
paths:
...
  /private/document-tags/tags/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет тег из системы и все его связи с документами
      parameters:
      - description: ID тега
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Тег успешно удален
        "400":
          description: Неверный формат ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Тег не найден
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удаление тега
      tags:
      - document-tags
    get:
      consumes:
      - application/json
      description: Возвращает информацию о теге по его идентификатору
      parameters:
      - description: ID тега
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Данные тега
          schema:
            $ref: '#/definitions/documenttag.TagResponse'
        "400":
          description: Неверный формат ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Тег не найден
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение тега по ID
      tags:
      - document-tags
    put:
      consumes:
      - application/json
      description: Изменяет название тега
      parameters:
      - description: ID тега
        in: path
        name: id
        required: true
        type: string
      - description: Новое название
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/documenttag.UpdateTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Тег успешно обновлен
          schema:
            $ref: '#/definitions/documenttag.TagResponse'
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Тег не найден
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновление тега
      tags:
      - document-tags
...
  /private/documents/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет документ и его файлы из хранилища
      parameters:
      - description: ID документа
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Документ успешно удален
        "400":
          description: Неверный формат ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Документ не найден
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удаление документа
      tags:
      - documents
    get:
      consumes:
      - application/json
      description: Возвращает информацию о документе с тегами и preview URL
      parameters:
      - description: ID документа
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Данные документа
          schema:
            $ref: '#/definitions/document.DocumentResponse'
        "400":
          description: Неверный формат ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Документ не найден
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение документа по ID
      tags:
      - documents
    put:
      consumes:
      - application/json
      description: Обновляет название, папку или отправителя документа
      parameters:
      - description: ID документа
        in: path
        name: id
        required: true
        type: string
      - description: Данные для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/document.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Документ успешно обновлен
          schema:
            $ref: '#/definitions/document.DocumentResponse'
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Документ не найден
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновление метаданных документа
      tags:
      - documents
...
  /private/documents/search:
    post:
      consumes:
      - application/json
      description: 'Ищет документы по различным критериям: текстовый запрос, папка,
        теги'
      parameters:
      - description: Параметры поиска
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/document.SearchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Результаты поиска
          schema:
            $ref: '#/definitions/document.DocumentsListResponse'
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Поиск документов
      tags:
      - documents
...
  /private/folders/{id}/rename:
    put:
      consumes:
      - application/json
      description: Изменяет название папки
      parameters:
      - description: ID папки
        in: path
        name: id
        required: true
        type: string
      - description: Новое название
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/folder.RenameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Папка успешно переименована
          schema:
            $ref: '#/definitions/folder.FolderResponse'
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Папка не найдена
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Переименование папки
      tags:
      - folders
  /private/folders/by-parent:
    post:
      consumes:
      - application/json
      description: Возвращает список папок по родительской папке. Если parent_id не
        указан, возвращает корневые папки
      parameters:
      - description: Параметры запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/folder.GetByParentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Список папок
          schema:
            $ref: '#/definitions/folder.FoldersListResponse'
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение вложенных папок
      tags:
      - folders
  /private/folders/company/{company_id}:
    get:
      consumes:
      - application/json
      description: Возвращает список всех папок компании без учета иерархии
      parameters:
      - description: ID компании
        in: path
        name: company_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список папок
          schema:
            $ref: '#/definitions/folder.FoldersListResponse'
        "400":
          description: Неверный формат ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение всех папок компании
      tags:
      - folders
...
swagger: "2.0"